修复缩放模式下，特效不正确的问题。改回hashstr的算法

+++ LONS/Scripter/LOScriptReader_command.cpp
@@ -544,7 +544,7 @@ int LOScriptReader::operationCommand(FunctionInterface *reader) {
 	//LOString gg[] = { "cos" ,"dec" ,"inc" ,"div" ,"mod" ,"mul" ,"rnd" ,"rnd2" ,"sin" ,"sub" ,"tan" };
 	//for (int ii = 0; ii < 11; ii++) printf("%s:%x\n",gg[ii].c_str(),  gg[ii].HashStr());
 
-	unsigned int optype = curCmd.HashStr();
+	unsigned int optype = (unsigned int)curCmd.HashStr();
 	int val = v1->GetReal();
 	switch (optype)
 	{
